split the huge formatter file
