--- src/formatter/format.rs
@@ -1,1 +1,0 @@
-use super::*;
